Update and rename StackDSCode.C to stack.C
